Bo sung flow

+++ app/src/main/java/com/example/vnhai/utils.kt
@@ -13,6 +13,10 @@ import java.io.InputStream
 
 var INTERNAL_STORAGE_DIR: String = ""
 
+enum class RemoteState{
+    Loading, Error, Success
+}
+
 fun String.onlyLetters() = all { it.isLetterOrDigit() }
 fun String.isValidateEmail() = endsWith("@apero.vn") && length > 9 &&
         subSequence(
@@ -32,7 +36,7 @@ fun getNameMusicFromPath(path: String): String
     return path.substring(path.lastIndexOf("/")+1, path.length)
 }
 
-fun saveFileToInternalStorage(filesDir: File, fileName: String, content: ResponseBody){
+fun saveFileToInternalStorage(filesDir: File, fileName: String, content: ResponseBody): String{
     val directory = File(filesDir, INTERNAL_STORAGE_DIR)
     if(!directory.exists()){
         directory.mkdir()
@@ -40,20 +44,23 @@ fun saveFileToInternalStorage(filesDir: File, fileName: String, content: Respons
 
     val inputStream: InputStream  = content.byteStream()
     val file = File(directory, fileName)
-    val fileOutputStream = FileOutputStream(file)
-    val buffer = ByteArray(8*1024)
-    var byteRead: Int
-
-    while(inputStream.read(buffer).also { byteRead = it } != -1)
+    if(!file.exists())
     {
-        fileOutputStream.write(buffer,0,byteRead)
+        val fileOutputStream = FileOutputStream(file)
+        val buffer = ByteArray(8*1024)
+        var byteRead: Int
+
+        while(inputStream.read(buffer).also { byteRead = it } != -1)
+        {
+            fileOutputStream.write(buffer,0,byteRead)
+        }
+        fileOutputStream.close()
     }
-    fileOutputStream.close()
+    return file.path
 }
 
 fun readFileFromInternalStorage(filesDir: String, filename: String){
     val directory = File(filesDir, INTERNAL_STORAGE_DIR)
-    Log.d("main", "utils $directory")
     val file = File(directory, filename)
 
     val fileInputStream = FileInputStream(file)
